fix: ci

+++ btc-types/src/header.rs
@@ -10,7 +10,7 @@ pub type Work = U256;
 
 pub const BLOCKS_PER_ADJUSTMENT: u64 = 2016;
 pub const TARGET_BLOCK_TIME_SECS: u64 = 10 * 60;
-pub const EXPECTED_TIME: u64 = BLOCKS_PER_ADJUSTMENT as u64 * TARGET_BLOCK_TIME_SECS;
+pub const EXPECTED_TIME: u64 = BLOCKS_PER_ADJUSTMENT * TARGET_BLOCK_TIME_SECS;
 pub const MAX_ADJUSTMENT_FACTOR: u64 = 4;
 pub const POW_LIMIT: U256 = U256::new(
     0x0000_0000_ffff_ffff_ffff_ffff_ffff_ffff,
